Monkey-patch SmallMover and test it
Also initialize PyRosetta with pyrosetta.init() on import.
The monkey-patched SmallMover was first intended as a child class
of the pyrosetta class, this way one would know which library is
used, and super() could be used to replace parent methods.
However it seems it clashes with some C++ virtual functions, which
don't work anymore and gives an error upon using assign method on a
pose. Monkey-patching with redefining every method ensues.